Add the GUI part.
Add the GUI part to the project, also modified some implementations.
Extract the ”paint” part back into main.hs, because users may want to
“paint” and see the graph without clicking “show graph”.
Fix the DOT node label problem (invalid symbol “|”).
Make the size of the node fixed, with “doublecircle” and “circle” look
the same size. But in some cases the circles are not big enough to hold
the labels, which remains as a problem.